+++ libs/ens/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.5.1](https://github.com/cowprotocol/cowswap/compare/ens-v3.5.0...ens-v3.5.1) (2026-08-12)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.13.0
+  - @cowprotocol/common-const@3.9.0
+
 ## [3.5.0](https://github.com/cowprotocol/cowswap/compare/ens-v3.4.7...ens-v3.5.0) (2026-08-05)
 
 ### Minor Changes